xdp-root-shim: Start between network.target and network-online.target.
Bringing up XDP on a new interface for the first time causes the interface to temporarily go down and back up. By ensuring that the XDP shim is brought online prior to `network-online.target`, we ensure that programs that rely on stable network interfaces are not started until this process has completed.